#12 filmorate. add-recommendations
#12 filmorate. add-recommendations

#12 filmorate. add-recommendations

#12 filmorate. финальный рефакторинг

#12 filmorate. финальный рефакторинг

+++ src/main/java/ru/yandex/practicum/filmorate/controllers/FilmController.java
@@ -67,7 +67,7 @@ public class FilmController {
     @GetMapping("/films/popular")
     public List<Film> getPopularFilmList(@RequestParam(defaultValue = "10", required = false) Integer count,
                                          @RequestParam(required = false) Integer year,
-                                         @RequestParam(required = false) Integer genreId) {
+                                         @RequestParam(required = false) Long genreId) {
         return filmService.getPopularFilmList(count, year, genreId);
     }
